+++ ts-definitions/tests.d.ts
@@ -469,6 +469,87 @@ declare namespace BrowserTests {
 		 * @param ctx
 		 */
 		isVisible(selectorOrElement: PlaywrightElContext | string, ctx?: PlaywrightElContext): Promise<boolean>;
+
+		/**
+		 * Returns a generator of element names
+		 *
+		 * @example
+		 * ```typescript
+		 * const
+		 *   base = elNameGenerator('p-index'), // Function
+		 *   elName = base('page'); // 'p-index__page'
+		 * ```
+		 */
+		elNameGenerator(blockName: string): (elName: string) => string;
+
+		/**
+		 * Returns a generator of element class names
+		 *
+		 * @example
+		 * ```typescript
+		 * const
+		 *   base = elNameSelectorGenerator('p-index'), // Function
+		 *   elName = base('page'); // '.p-index__page'
+		 * ```
+		 */
+		elNameSelectorGenerator(blockName: string): (elName: string) => string;
+
+		/**
+		 * Returns a generator of element names with modifiers
+		 *
+		 * @example
+		 * ```typescript
+		 * const
+		 *   base = elNameGenerator('p-index') // Function,
+		 *   elName = base('page'), // 'p-index__page'
+		 *   modsBase = elModNameGenerator(elName), // Function
+		 *   elNameWithMods = modsBase('type', 'test'); // 'p-index__page_type_test'
+		 * ```
+		 *
+		 * @example
+		 * ```typescript
+		 * const
+		 *   base = elNameGenerator('p-index') // Function,
+		 *   elName = base('page'), // 'p-index__page'
+		 *   modsBase = elModNameGenerator(elName, 'type', 'test'); // 'p-index__page_type_test'
+		 * ```
+		 */
+		elModNameGenerator(fullElName: string): (modName: string, modVal: string) => string;
+		elModNameGenerator(fullElName: string, modName: string, modVal: string): string;
+		elModNameGenerator(
+			fullElName: string,
+			modName?: string,
+			modVal?: string
+		): ((modName: string, modVal: string) => string) | string;
+
+		/**
+		 * Returns a generator of element names with modifiers
+		 *
+		 * @example
+		 * ```typescript
+		 * const
+		 *   base = elNameGenerator('p-index') // Function,
+		 *   elName = base('page'), // 'p-index__page'
+		 *   modsBase = elModNameGenerator(elName), // Function
+		 *   elNameWithMods = modsBase('type', 'test'); // '.p-index__page_type_test'
+		 * ```
+		 *
+		 * @example
+		 * ```typescript
+		 * const
+		 *   base = elNameGenerator('p-index') // Function,
+		 *   elName = base('page'), // 'p-index__page'
+		 *   modsBase = elModSelectorGenerator(elName, 'type', 'test'); // '.p-index__page_type_test'
+		 * ```
+		 */
+		elModSelectorGenerator(fullElName: string): (modName: string, modVal: string) => string;
+		elModSelectorGenerator(fullElName: string, modName: string, modVal: string): string;
+		elModSelectorGenerator(
+			fullElName: string,
+			modName?: string,
+			modVal?: string
+		): ((modName: string, modVal: string) => string) | string;
+
 	}
 
 	/**