+++ CHANGELOG.md
@@ -11,6 +11,14 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
+## v3.0.0-rc.86 (2020-10-12)
+
+* Added new methods `tests/helpers/dom`
+  * `elNameGenerator`
+  * `elNameSelectorGenerator`
+  * `elModNameGenerator`
+  * `elModSelectorGenerator`
+
 ## v3.0.0-rc.85 (2020-10-09)
 
 #### :boom: Breaking Change